bot/requirements.txt
discord.py==2.4.0
SQLAlchemy==2.0.34
aiohttp==3.10.5